requirements.txt
certifi==2016.2.28
numpy==1.13.1
olefile==0.44
Pillow==4.3.0
